// src/app/models/reports/ORMYearlyCollectionTotal.ts
export class ORMYearlyCollectionTotal{
    location_id: String;
    location_name: String;
    collection_type: String;
    month_01: Number=0;
    month_02: Number=0;
    month_03: Number=0;
    month_04: Number=0;
    month_05: Number=0;
    month_06: Number=0;
    month_07: Number=0;
    month_08: Number=0;
    month_09: Number=0;
    month_10: Number=0;
    month_11: Number=0;
    month_12: Number=0;
    total: Number=0;
}